refactor PluginModal: move each tab into its own component
