+++ workbench/src/renderer/styles/style.css
@@ -1005,7 +1005,7 @@ code {
   border-bottom: 1px solid rgba(0,0,0,.125);
   border-radius: 0;
   margin-bottom: 0;
-  padding: 0.5rem;;
+  padding: 0.5rem;
   &.plugin-registry-button {
     display: flex;
     align-items: center;
@@ -1028,6 +1028,8 @@ code {
 }
 
 .registry-install-form {
+  padding-top: 1rem;
+  padding-bottom: 1rem;
   & form > div {
     & > div, & > button {
       margin-top: 1rem;
@@ -1061,4 +1063,12 @@ code {
   width: 3em;
   margin-bottom: 1rem;
   color: var(--digital-red);
+}
+
+.installed-plugin-row {
+  border-top: 1px solid var(--bs-modal-header-border-color);
+}
+
+.installed-plugin-source-info {
+  font-size: 0.875rem;
 }
